Added login integration and responder report disaster
All the endpoints can now be called from functions in the api folder. Its broken down the same way as the routes folder in the backend. Thus, all you need to do is import one of those functions.

Those api functions may been adjustments depending on the desired format of the output.

I am working on the CreateDisaster page but I am also change the api folder to be classes. Will see.