fixed errors and improve functionality
To Do:
Better communicate to the user when there is insufficient capacity.
Improve the orders table to have sorting and filtering.
Fix the alert on the disaster information page when you first load to it.
Add the orders/:id page with the map functionality or substitute the general page for it in the frontend and backend.
Think about the evacuation checkbox in the disaster information page (its a bit confusing).
Make sure the alert comes up when add report fails (or a way to communicate a successful submission).